% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGravyScore.R
\name{calculateGravyScore}
\alias{calculateGravyScore}
\title{Calculate Gravy scores of an amino acid/protein}
\usage{
calculateGravyScore(aa)
}
\arguments{
\item{aa}{\code{AAString} containing the amino acid, peptide, or 
protein sequence}
}
\description{
The function \code{calculateGravyScore} calculates the GRAVY (grand average 
of hydropathy) value of a given amino acid/protein sequence. 
The GRAVY value is calculated by adding the hydropathy value for each 
residue and dividing by the length of the sequence.
}
\details{
The hydropathy values are taken from Table 2 in Kyte and Doolittle (1982).
}
\examples{
library(Biostrings)
aa <- AAString(x = "TEST")
calculateGravyScore(aa = aa)
}
\references{
Kyte and Doolittle (1982):
'A Simple Method for Displaying the  Hydropathic Character of a Protein',
Journal of Molecular Biology, 175, 105-132.
doi: 10.1016/0022-2836(82)90515-0
}
