% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIsoelectricPoint.R
\name{calculateIsoelectricPoint}
\alias{calculateIsoelectricPoint}
\title{Calculate isoelectric point of an amino acid/protein}
\usage{
calculateIsoelectricPoint(
  aa,
  method = c("EMBOSS", "DTASelect", "Solomon", "Sillero", "Rodwell", "Lehninger",
    "Toseland", "Thurlkill", "Nozaki", "IPC_protein", "IPC_peptide")
)
}
\arguments{
\item{aa}{`character(1)`, vector containing the amino acid, peptide, or 
protein sequence}

\item{method}{`character(1)`, vector specifying the method/parametrization
for the pKa values}
}
\description{
The function `calculateIsoelectricPoint` calculates the isoelectric point 
of a given amino acid/protein sequence. 
Several methods are implemented that use different pKa values for the 
ionizable groups of proteins (N terminal, C terminal, and amino acids
C, D, E, H, K, R, and Y).
}
\details{
The pKa values are taken from Table 4 in Kozlowski (2016).

For polypeptides, the isoelectric point depends primarily on the dissociation 
constants (pKa) for the ionizable groups of seven charged amino acids: 
glutamate (\\delta-carboxyl group), aspartate (\\beta-carboxyl group), cysteine 
(thiol group), tyrosine (phenol group), histidine (imidazole side chains), 
lysine (\\epsilon-ammonium group) and arginine (guanidinium group). Moreover, 
the charge of the terminal groups (NH2 and COOH) can greatly affect the pI 
of short peptides. Generally, the Glu, Asp, Cys, and Tyr ionizable groups 
are uncharged below their pKa and negatively charged above their pKa. 
Similarly, the His, Lys, and Arg ionizable groups are positively charged 
below their pKa and uncharged above their pKa.

The implemented algorithm (bisection algorithm) is as in Kozlowski (2016).
}
\examples{
aa <- "TEST"
calculateIsoelectricPoint(aa = aa, method = "EMBOSS")
}
\references{
Kozlowski (2016): 'IPC - Isoelectric Point Calculator', Biology Direct, 
11, 55.
doi: doi.org/10.1186/s13062-016-0159-9
}
