% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_checkAA.R
\name{checkAA}
\alias{checkAA}
\title{Check if amino acid sequence contains right characters}
\usage{
checkAA(aa)
}
\arguments{
\item{aa}{\code{AAString} containing the amino acid, peptide, or protein 
sequence}
}
\value{
character
}
\description{
The function \code{checkAA} checks if the letters in \code{aa} contain the 
correct characters (all letters in \code{aa} have to be in 
\code{AA_ALPHABET}). The function will return the sequence as a 
character vector or will break otherwise.
}
\details{
\code{aa} has to be an \code{AAString} object.
}
\examples{
library(Biostrings)
aa <- AAString(x = "TEST")
PhysicoChemicalPropertiesProtein:::checkAA(aa)
}
