% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydropathy-data.R
\docType{data}
\name{hydropathy}
\alias{hydropathy}
\title{Hydropathy values for amino acids}
\format{
\code{data.frame}
}
\source{
hydropathy <- rbind(
    c("I", "Ile", "isoleucine", 4.5),
    c("V", "Val", "valine", 4.2),
    c("L", "Leu", "leucine", 3.8),
    c("F", "Phe", "phenylalanine", 2.8),
    c("C", "Cys", "cysteine", 2.5),
    c("M", "Met", "methionine", 1.9),
    c("A", "Ala", "alanine", 1.8),
    c("G", "Gly", "glycine", -0.4),
    c("T", "Thr", "threonine", -0.7),
    c("W", "Trp", "tryptophan", -0.9),
    c("S", "Ser", "serine", -0.8),
    c("Y", "Tyr", "tyrosine", -1.3),
    c("P", "Pro", "proline", -1.6),
    c("H", "His", "histidine", -3.2),
    c("E", "Glu", "glutamic acid", -3.5),
    c("Q", "Gln", "glutamine", -3.5),
    c("D", "Asp", "aspartic acid", -3.5),
    c("N", "Asn", "asparagine", -3.5),
    c("K", "Lys", "lysine", -3.9),
    c("R", "Arg", "arginine", -4.5))
    
hydropathy <- as.data.frame(hydropathy)
colnames(hydropathy) <- c("SINGLE_LETTER", "THREE_LETTER", "NAME", "SCORE")
rownames(hydropathy) <- hydropathy$SINGLE_LETTER
hydropathy$SCORE <- as.numeric(hydropathy$SCORE)

saveRDS(hydropathy, "hydropathy.RDS")
}
\value{
\code{data.frame}
}
\description{
\code{hydropathy} contains hydropathy values for amino acid residues.
}
\details{
The hydropathy values are taken from Table 2 in Kyte and Doolittle (1982).
}
\references{
Kyte and Doolittle (1982):
'A Simple Method for Displaying the  Hydropathic Character of a Protein',
Journal of Molecular Biology, 175, 105-132.
doi: 10.1016/0022-2836(82)90515-0
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
