\name{physicoChemicalPropertiesProtein-package}
\alias{physicoChemicalPropertiesProtein-package}
\alias{physicoChemicalPropertiesProtein}
\docType{package}
\title{
\packageTitle{physicoChemicalPropertiesProtein}
}
\description{
\packageDescription{physicoChemicalPropertiesProtein}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{physicoChemicalPropertiesProtein}
\packageIndices{physicoChemicalPropertiesProtein}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{physicoChemicalPropertiesProtein}

Maintainer: \packageMaintainer{physicoChemicalPropertiesProtein}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
